Create HoldState
Created a new hold state that uses the intake lidars to figure out
whether it is holding a cone or a cube, and runs the correct grab method
to keep a hold of the cone or cube.
Removed the other hold states because they are no longer necessary.
